+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- sectionLinks now contains the prop `target` in links where is possible to use one of the values `[self_tab, new_tab]`. `self_tab` it means the value for target attribute it will be `_self`. `new_tab` it will be `_blank`
+- sectionLinks now contains the prop `target` in links where is possible to use one of the values `[_self_, _blank_]`. That will define the value for target attribute. `<a href="/" target="_self">`
 
 ## [2.6.13] - 2019-03-14
 ### Changed